ajuste simulacao para sqlLite

+++ src/main/java/br/com/mattos/simuladorinvestimento/application/dto/request/SimularInvestimentoRequestDTO.java
@@ -4,6 +4,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
-public record SimularInvestimentoRequestDTO(@NotNull @Positive Long clienteId, @NotNull @Positive Double valor, @NotNull @Positive Integer prazoMeses,@NotBlank String tipoProduto) {
+public record SimularInvestimentoRequestDTO(@NotNull @Positive Integer clienteId, @NotNull @Positive Double valor,
+                                            @NotNull @Positive Integer prazoMeses,@NotBlank String tipoProduto) {
 
 }
