+++ src/main/java/br/com/mattos/simuladorinvestimento/application/dto/response/ProdutoSimulacaoResponseDTO.java
@@ -1,6 +1,6 @@
 package br.com.mattos.simuladorinvestimento.application.dto.response;
 
-public record ProdutoSimulacaoResponseDTO(Long id, String nome, String tipo, Double rentabilidade, String risco) {
+public record ProdutoSimulacaoResponseDTO(Integer id, String nome, String tipo, Double rentabilidade, String risco) {
 
 }
