+++ src/main/java/br/com/mattos/simuladorinvestimento/domain/model/SimulacaoInvestimento.java
@@ -2,8 +2,8 @@ package br.com.mattos.simuladorinvestimento.domain.model;
 
 import java.time.Instant;
 
-public record SimulacaoInvestimento(Long clientId, Produto produto, ResultadoSimulacao resultado, Instant dataSimulacao) {
-    public SimulacaoInvestimento(Long clientId, Produto produto, ResultadoSimulacao resultado) {
+public record SimulacaoInvestimento(Integer clientId, Produto produto, ResultadoSimulacao resultado, Instant dataSimulacao) {
+    public SimulacaoInvestimento(Integer clientId, Produto produto, ResultadoSimulacao resultado) {
         this(clientId, produto, resultado, Instant.now());
     }
 }
